Refactor Filter Control Select input initialization
Extract the conditions in which the Select dropdown for
the input type in the filter control is initialized. Moved
this into its own function (based on the single responsibility
principle) and in general made the code more reader friendly.

The formatted field value will also appear as a formatted field
value (before it was unclear what the relationship between value
and text in addOptionToSelectControl was).