refactor filter cookies extension to avoid dbcalls
There was an incompatibility between the cookies extension and the
filter control extension, which resulted in the following behavior:
every time that a cookie was found, it would be added to a list of
filter, and it would query the database again. This would result in (for
example) a series of three filters to force three separate database
calls, each database call accumulating the last.

This update fixes this issue. it also incorporates the change from issue
 #1605 -- in which updatePagination is called by onSearch(), but both
are included in the onColumnSearch() function, causing the table to
double all calls to the database.