Fix cookie extension onPage*** hook return values
The cookie extension hooks some of the bootstrap table methods to
implement saving the state. Unfortunately the onPage*** hooks don't
propagate the return values of the original hooked methods.

The original methods return false to ask jQuery to call
event.preventDefault() and event.stopPropagation(). Without this
clicking the pagination control caused my webpage to scroll to the top.